Removed cache, refresh of add friends, tidy up of warnings
Removed the caching as this should never be done on the api, this was causing a lot of problems. Forced the fragment to refresh when you add a friend or request to add a friend. Cleaned up a few easy warnings such as unused variables and the unused argument of the UserViewHolder.kt.