Fixed add vote plan to calendar button; Fixed get vote reminders api call; Added toast messages for error handling.
Temporary hardcoded vote reminders has been added whilst backend issues
are getting sorted.

There was a special case to return null for paginated api calls with 0
records, this was the cause of this bug and had no clear purpose so was
removed.

Also, changed how toasts are created in friends fragment for wider
compatibility.